+++ app/scripts/HiGlassComponent.js
@@ -715,7 +715,6 @@ class HiGlassComponent extends React.Component {
    * visible?
    */
   isEditable() {
-    // console.log('editable:', this.props.options, this.state.viewConfig);
     if (!this.props.options || !('editable' in this.props.options)) {
       return this.state.viewConfig.editable;
     }
@@ -728,14 +727,43 @@ class HiGlassComponent extends React.Component {
    * visible?
    */
   isTrackMenuDisabled() {
-    const fromOptions = this.props.options && this.props.options.disableTrackMenu;
-    const fromViewconf = this.state.viewConfig && this.state.viewConfig.disableTrackMenu;
-
-    if (!this.props.options || !('disableTrackMenu' in this.props.options)) {
-      return fromViewconf;
+    if (
+      this.props.options && (
+        this.props.options.editable === false
+        || this.props.options.tracksEditable === false
+      )
+    ) {
+      return true;
     }
 
-    return fromOptions || fromViewconf;
+    return (
+      this.state.viewConfig && (
+        this.state.viewConfig.tracksEditable === false
+        || this.state.viewConfig.editable === false
+      )
+    );
+  }
+
+  /**
+   * Can views be added, removed or rearranged and are the view headers
+   * visible?
+   */
+  isViewHeaderDisabled() {
+    if (
+      this.props.options && (
+        this.props.options.editable === false
+        || this.props.options.viewEditable === false
+      )
+    ) {
+      return true;
+    }
+
+    return (
+      this.state.viewConfig && (
+        this.state.viewConfig.viewEditable === false
+        || this.state.viewConfig.editable === false
+      )
+    );
   }
 
   /**
@@ -3663,7 +3691,11 @@ class HiGlassComponent extends React.Component {
           );
         };
 
-        const multiTrackHeader = this.isEditable() && !this.state.viewConfig.hideHeader ? (
+        const multiTrackHeader = (
+          this.isEditable()
+          && !this.isViewHeaderDisabled()
+          && !this.state.viewConfig.hideHeader
+        ) ? (
           <ViewHeader
             ref={(c) => { this.viewHeaders[view.uid] = c; }}
             getGenomePositionSearchBox={getGenomePositionSearchBox}
@@ -3715,7 +3747,7 @@ class HiGlassComponent extends React.Component {
             onZoomToData={uid => this.handleZoomToData(uid)}
             viewUid={view.uid}
           />
-        ) : null;
+          ) : null;
 
         return (
           <div